Update .env variables names and values
The commit updates several .env variable names and values for a clearer and more accurate configuration. The changes include renaming CLIENT_ID and CLIENT_SECRET to inline with Application Properties. The flag to determine if secret is base64 is also added. In addition, a boolean flag is utilized to enable the use of Auth0 IDP Provider instead of setting the provider's name directly.